Refactor pipeline to use Kyaml
This commit refactors the pipeline to be based on the Kyaml pipeline
concept. Instead of manually maintaining the list of functions, and
iterating over them - we translate the Kude structures into the Kyaml
pipeline structure and execute it.

This allows us to build on the stability and features of the Kyaml
library, while still delivering our unique features.