Add "helm" function
This change adds the "helm" generator function, which can execute Helm
with arbitrary arguments and pipe its output as YAML resources as the
function output.

Naturally, the only use-case here that makes sense is to invoke Helm
with the "template" command - as that's the only Helm command that
outputs YAML (the others actually install things on the target cluster),
 but we decided to none-the-less keep this function generic and allow
 the caller to invoke arbitrary arguments to Helm.

The function will download the appropriate Helm executable based on the
given Helm version in the configuration (with a default of 3.8.1). The
Helm executable will be cached on the host machine for future
invocations.