Allow user to control network access for functions
This change allows the user to control whether certain functions are
allowed network access or not. Some functions will require network
access, but the ultimate decision of whether to allow it must be
reserved for the user.

This change allows the user to specify the "network" boolean key under
each function, to decide whether that function should receive network
access or not (default is to disable network).