Sign the macOS binary
Update the CI process to sign the macOS executable binary, to allow
macOS users to run it without bypassing Gatekeeper. The binary is signed
with my personal Apple Developer ID.

This change also changes the OS for the GitHub runner that generates the
macOS binary from Ubuntu to macOS. This unfortunately means, however,
that running tests is no longer possible in that runner - since Docker
is not yet available for GitHub Actions workflows running on macOS (for
details, see [1]).

[1] https://github.community/t/is-it-possible-to-install-and-configure-docker-on-macos-runner/16981/11