Fix file mounts of nested Kude packages (#56)
The change fixes the base directory against which local file mount paths
are resolved in nested imported Kude packages. For example, imagine a
Kude pckage "a", which imports another Kude package "b"; and package "b"
mounts a file called "./f.txt" - it will be resolved against the base
dir of package "a", instead of package "b".