Rename configmap & secret functions, refactor tests
This change refactors the testing harness, but switching to a code
generation approach: running `go generate` at the root of the project
will generate Go test cases for every directory where both `kude.yaml` &
`expected.yaml` are found.

The idea is that each such pair represents a test. It will be placed in
a `scenario_test.go` file placed in that directory, generated from a
Go template. Thanks to this change, some tests are moved to the relevant
function directory that they were intended to test. Some tests, however,
were retained in the "test/scenarios" directory as they are testing a
more generic feature than just a specific function.

This change also renames the "configmap" and "secret" functions, by
adding a "create-" prefix to both of them. Therefore:
* "configmap" becomes "create-configmap"
* "secret" becomes "create-secret"

This adds more clarity to each function's purpose.